update_toj261

+++ toj/Pro261 陰謀論.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 int main()
 {
-	int a, b;
+	long double a, b;
 	cin >> a >> b;
 	if ( a==b && a>0 )
 		cout << "Nymphia" << endl;
